refactor: user 필드 수정
