+++ src/main/java/com/example/ensurify/repository/UserMeetingRoomRepository.java
@@ -1,6 +1,6 @@
 package com.example.ensurify.repository;
 
-import com.example.ensurify.domain.UserMeetingRoom;
+import com.example.ensurify.domain.mappings.UserMeetingRoom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserMeetingRoomRepository extends JpaRepository<UserMeetingRoom, Long> {
