+++ src/main/java/com/example/ensurify/service/MeetingRoomService.java
@@ -4,8 +4,7 @@ import com.example.ensurify.common.apiPayload.code.status.ErrorStatus;
 import com.example.ensurify.common.apiPayload.exception.GeneralException;
 import com.example.ensurify.domain.*;
 import com.example.ensurify.domain.enums.Role;
-import com.example.ensurify.dto.request.CreateContractRequest;
-import com.example.ensurify.dto.response.CreateContractResponse;
+import com.example.ensurify.domain.mappings.UserMeetingRoom;
 import com.example.ensurify.repository.MeetingRoomRepository;
 import com.example.ensurify.repository.UserMeetingRoomRepository;
 import lombok.RequiredArgsConstructor;
